feat(FileUploader): remove old file uploader init new upload
