+++ tsconfig.json
@@ -11,7 +11,7 @@
     "lib": ["ES2021"],
     "paths": {
       "@tiptap-location/*": ["./tiptap/*"],
-      "@/*": ["@/*"]
+      "@/*": ["@/*", "./src/*"]
     }
   }
 }
